requirements.txt
aioredis==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
conda==4.3.16
databases==0.5.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
filelock==3.6.0
greenlet==1.1.2
idna==3.3
mypy==0.931
mypy-extensions==0.4.3
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycosat==0.6.3
pycparser==2.21
pydantic==1.8.2
pydbantic==0.0.22
python-dotenv==0.19.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SQLAlchemy==1.4.28
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.2
virtualenv-clone==0.5.7
